telegram-bot/requirements.txt
python-telegram-bot==20.6
aiohttp>=3.8.0
python-dateutil>=2.8.2
